requirements.txt
# Production Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
langchain>=0.1.6
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
openai>=1.12.0
anthropic>=0.18.0
unstructured[local-inference,pdf]>=0.11.8
python-multipart>=0.0.9
python-dotenv>=1.0.1
serpapi>=0.1.5
tiktoken>=0.6.0

# Development Dependencies
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.1.1
ruff>=0.1.12

